tests: rename cql_types.rs -> serialization.rs
